Break out by middleware, not handler
