+++ src/bindings.d.ts
@@ -1,6 +1,6 @@
 import { CID } from '@web3-storage/gateway-lib/handlers'
-import { Environment as RateLimiterEnvironment } from './handlers/rate-limiter.types.ts'
-import { Environment as CarBlockEnvironment } from './handlers/car-block.types.ts'
+import { Environment as RateLimiterEnvironment } from './middleware/withRateLimit.types.ts'
+import { Environment as CarBlockEnvironment } from './middleware/withCarBlockHandler.types.ts'
 
 export interface Environment extends CarBlockEnvironment, RateLimiterEnvironment {
   VERSION: string
