+++ src/middleware/withRateLimit.types.ts
@@ -1,7 +1,7 @@
 import { CID } from '@web3-storage/gateway-lib/handlers'
 import { Environment as MiddlewareEnvironment } from '@web3-storage/gateway-lib'
 import { KVNamespace, RateLimit } from '@cloudflare/workers-types'
-import { RATE_LIMIT_EXCEEDED } from './../constants.js'
+import { RATE_LIMIT_EXCEEDED } from '../constants.js'
 
 export interface Environment extends MiddlewareEnvironment {
   ACCOUNTING_SERVICE_URL: string
